BUG-31942 Implement unit tests for RemoteBlockReader
